# perf: improve performance of serializing log timestamps