--- a/internals/servicelog/formatter_benchmark_test.go
+++ b/internals/servicelog/formatter_benchmark_test.go
@@ -19,6 +19,17 @@ import (
 	"time"
 )
 
+func BenchmarkAppendFormat(b *testing.B) {
+	t := time.Now()
+	buf := make([]byte, 0, 64)
+
+	for i := 0; i < b.N; i++ {
+		buf = t.UTC().AppendFormat(buf[:0], "2006-01-02T15:04:05.000Z")
+	}
+
+	_ = buf // ensure buf is not optimized away
+}
+
 func BenchmarkAppendTimestamp(b *testing.B) {
 	t := time.Now()
 	buf := make([]byte, 0, 64)