--- a/internals/servicelog/formatter_test.go
+++ b/internals/servicelog/formatter_test.go
@@ -12,15 +12,14 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>
 
-package servicelog_test
+package servicelog
 
 import (
 	"bytes"
 	"fmt"
+	"time"
 
 	. "gopkg.in/check.v1"
-
-	"github.com/canonical/pebble/internals/servicelog"
 )
 
 const (
@@ -33,7 +32,7 @@ var _ = Suite(&formatterSuite{})
 
 func (s *formatterSuite) TestFormat(c *C) {
 	b := &bytes.Buffer{}
-	w := servicelog.NewFormatWriter(b, "test")
+	w := NewFormatWriter(b, "test")
 
 	fmt.Fprintln(w, "first")
 	fmt.Fprintln(w, "second")
@@ -48,7 +47,7 @@ func (s *formatterSuite) TestFormat(c *C) {
 
 func (s *formatterSuite) TestFormatSingleWrite(c *C) {
 	b := &bytes.Buffer{}
-	w := servicelog.NewFormatWriter(b, "test")
+	w := NewFormatWriter(b, "test")
 
 	fmt.Fprintf(w, "first\nsecond\nthird\n")
 
@@ -57,4 +56,19 @@ func (s *formatterSuite) TestFormatSingleWrite(c *C) {
 %[1]s \[test\] second
 %[1]s \[test\] third
 `[1:], timeFormatRegex))
+}
+
+func (s *formatterSuite) TestAppendTimestamp(c *C) {
+	now := time.Now()
+	c.Assert(string(appendTimestamp(nil, now)), Equals,
+		now.UTC().Format("2006-01-02T15:04:05.000Z"))
+
+	c.Assert(string(appendTimestamp(nil, time.Time{})), Equals,
+		"0001-01-01T00:00:00.000Z")
+	c.Assert(string(appendTimestamp(nil, time.Date(2042, 12, 31, 23, 59, 48, 123_456_789, time.UTC))), Equals,
+		"2042-12-31T23:59:48.123Z")
+	c.Assert(string(appendTimestamp(nil, time.Date(2025, 8, 9, 1, 2, 3, 4_000_000, time.UTC))), Equals,
+		"2025-08-09T01:02:03.004Z")
+	c.Assert(string(appendTimestamp(nil, time.Date(2025, 8, 9, 1, 2, 3, 4_999_999, time.UTC))), Equals,
+		"2025-08-09T01:02:03.004Z") // time.Format truncates (not rounds) milliseconds too
 }