Because the timestamp format used is not exactly RFC3339, the Go
standard library falls back to generic formatting code that analyzes the
format on every append, this causes a lot of unnecessary work. Using a
specialized formatter specifically for the format, a ~72% improvement
is achieved.

```
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: github.com/canonical/pebble/internals/servicelog
cpu: Apple M1 Max
                   │   old.txt    │               new.txt               │
                   │    sec/op    │   sec/op     vs base                │
AppendTimestamp-10   120.55n ± 0%   33.27n ± 1%  -72.41% (p=0.000 n=10)
```

After this, the only improvement that could still be made is assuming that it's always UTC, which we already know, profililing data shows that could at most improve another ~12% (`time.(*Location).get`).

https://pprof.me/da1dac7424551de8264abc96c430f319/?profileType=profile%3Acpu%3Ananoseconds%3Acpu%3Ananoseconds%3Adelta&color_by=filename&dashboard_items=flamegraph%2Ctable&table_columns=color%2Cflat%2CflatPercentage%2Ccumulative%2Cname