@benhoyt 2025-08-11
Thanks for this @brancz. I've looked over it and done some further tests. I get a similar result on my amd64 machine (-67.35% vs base). I've also tweaked some comments, used `slices.Grow` (new in Go 1.21) instead of a hand-rolled version, and added some more explicit tests.